Textbook-The Hardware of a Quantum Computer
How all the building blocks of a quantum computer work, and deeper information on the qubits which lie at the heart of a quantum computer and internet. How qubits can be used and controlled efficiently, and the workings of the four most promising types of solid-state qubits: Silicon Spin Qubit, Diamond NV Center Qubit, Superconducting Transmon Qubit and Topological Qubit.